set bloomfilter size to 64_bit

+++ src/sorter/distributed/bloomfilter.hpp
@@ -1039,7 +1039,7 @@ public:
     BloomFilter(const size_t size, const size_t startDepth)
         : size(size), hashValues(size, 0), startDepth(startDepth) {}
     //~BloomFilter() { delete[] hashValues; }
-    const size_t bloomFilterSize = std::numeric_limits<uint32_t>::
+    const size_t bloomFilterSize = std::numeric_limits<uint64_t>::
         max(); // set to this size because distribution/load balancing was not
                // good enough TODO Discuss Multisequence Selection?
